RAG-Based-PDF-Chat=Bot-with-Agent-SDK/RAG-Based-PDF-Chat-Bot-with-Agent-SDK/requirements.txt
# Core
streamlit
pandas
numpy
torch
python-dotenv
langchain-core
unstructured
lxml
python-magic
langchain
langchain-community
langchain-openai
langchain_tavily
langgraph
openai
openai-agents
tiktoken
faiss-cpu
sentence-transformers
pydantic
pdfplumber
PyMuPDF
pytesseract
pillow
camelot-py[cv]
tabula-py

# UI
# gradio

# Voice (optional; install later if needed)
# openai-whisper
# pyttsx3
